Create package for movement related code and move it into there.
