+++ src/main/java/org/terasology/gooeyDefence/EnemyManager.java
@@ -23,17 +23,18 @@ import org.terasology.entitySystem.entity.EntityRef;
 import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterSystem;
-import org.terasology.gooeyDefence.components.enemies.BlankPathComponent;
-import org.terasology.gooeyDefence.components.enemies.EntrancePathComponent;
+import org.terasology.gooeyDefence.movement.components.BlankPathComponent;
+import org.terasology.gooeyDefence.movement.components.EntrancePathComponent;
 import org.terasology.gooeyDefence.components.enemies.GooeyComponent;
-import org.terasology.gooeyDefence.components.enemies.MovementComponent;
-import org.terasology.gooeyDefence.components.enemies.PathComponent;
+import org.terasology.gooeyDefence.movement.components.MovementComponent;
+import org.terasology.gooeyDefence.movement.components.PathComponent;
 import org.terasology.gooeyDefence.events.OnEntrancePathChanged;
 import org.terasology.gooeyDefence.events.OnFieldActivated;
-import org.terasology.gooeyDefence.events.ReachedGoalEvent;
-import org.terasology.gooeyDefence.events.RepathEnemyRequest;
+import org.terasology.gooeyDefence.movement.events.ReachedGoalEvent;
+import org.terasology.gooeyDefence.movement.events.RepathEnemyRequest;
 import org.terasology.gooeyDefence.events.health.DamageEntityEvent;
 import org.terasology.gooeyDefence.events.health.EntityDeathEvent;
+import org.terasology.gooeyDefence.movement.PathfindingManager;
 import org.terasology.logic.common.ActivateEvent;
 import org.terasology.logic.delay.DelayManager;
 import org.terasology.logic.location.LocationComponent;
