+++ src/main/java/org/terasology/gooeyDefence/movement/PathfindingManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 MovingBlocks
+ * Copyright 2018 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.gooeyDefence;
+package org.terasology.gooeyDefence.movement;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,12 +23,12 @@ import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterSystem;
 import org.terasology.flexiblepathfinding.JPSConfig;
 import org.terasology.flexiblepathfinding.PathfinderSystem;
-import org.terasology.gooeyDefence.components.enemies.BlankPathComponent;
-import org.terasology.gooeyDefence.components.enemies.CustomPathComponent;
+import org.terasology.gooeyDefence.DefenceField;
 import org.terasology.gooeyDefence.events.OnEntrancePathChanged;
 import org.terasology.gooeyDefence.events.OnFieldActivated;
-import org.terasology.gooeyDefence.events.RepathEnemyRequest;
-import org.terasology.gooeyDefence.pathfinding.EnemyWalkingPlugin;
+import org.terasology.gooeyDefence.movement.components.BlankPathComponent;
+import org.terasology.gooeyDefence.movement.components.CustomPathComponent;
+import org.terasology.gooeyDefence.movement.events.RepathEnemyRequest;
 import org.terasology.logic.location.LocationComponent;
 import org.terasology.math.geom.Vector3i;
 import org.terasology.registry.In;
