+++ src/main/java/org/terasology/gooeyDefence/movement/events/ReachedGoalEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.gooeyDefence.events;
+package org.terasology.gooeyDefence.movement.events;
 
 import org.terasology.entitySystem.event.ConsumableEvent;
