+++ src/main/java/org/terasology/gooeyDefence/towerBlocks/effectors/StunEffectorSystem.java
@@ -20,9 +20,9 @@ import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterSystem;
 import org.terasology.gooeyDefence.DefenceField;
-import org.terasology.gooeyDefence.components.enemies.BlankPathComponent;
+import org.terasology.gooeyDefence.movement.components.BlankPathComponent;
 import org.terasology.gooeyDefence.components.enemies.GooeyComponent;
-import org.terasology.gooeyDefence.components.enemies.PathComponent;
+import org.terasology.gooeyDefence.movement.components.PathComponent;
 import org.terasology.gooeyDefence.events.combat.ApplyEffectEvent;
 import org.terasology.logic.delay.DelayManager;
 import org.terasology.logic.delay.DelayedActionTriggeredEvent;
